scripts/workflows/fstests/copy-results.sh: use kdevops-results-archive
The kdevops-results-archive [0] is to be used with kdevops for archiving
results. Ensure that is done. Do sanity checks too, so we don't move
forward unless the user has results already and the symlink is created,
if the symlink to the results is not there, we create it for you. The
directory we check for the results archive is based off of the parent
current directory from where you run the script from, we check there for
the kdevops-results-archive. If found we create the symlink for you.

Each workflow can use its own symlink.

[0] https://github.com/linux-kdevops/kdevops-results-archive

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>