.gitignore: avoid including results
The results directory is now completely ephemeral. The archive is all
we used to use before but that is part of a new git tree, which we
will add support for next.

Ensure no one adds any more results as experience shows we want to keep
this repo only with code and expunges for the known baseline for kernels
that is all. Results must be dealt with separately.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>