defconfig: add LBS defconfigs for XFS
Add two LBS configs which demonstrates how to test LBS today,
you have the full set of XFS profiles to use and then a minimal
setup. For both we enable building Linux using Linus' latest rc
release and we then use the message ID for the latest series for
the LBS patches.

Which option you use depends on your interest and testing resources
available:

  a) make defconfig-lbs-small   enables only 2 XFS profiles to test
  b) make defconfig-lbs         enables all 28 XFS profiles to test

Use a) to try things out first, and then b) once you know what profiles
you want to test, you'd just disable the ones you don't want.

The point of this commit is not only to enable to easily test LBS
but also to showcase how easy it is now to test upstream patches
with kdevops constructs. As the LBS patches move forward all we have
to do is change the message ID and perhaps the base kernel used and
associated base tag.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>