common: move getting user and group inference to shared role
It is often that some ansible roles need to use an inferred user.
It used to be that our default username and group was "vagrant"
but with guestfs that is now "kdevops". So adjust the ordering
and special-case vagrant now.

This let's us include the common role when we need the inferred user
on any other task in another ansible role. We'll use it next.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>