fstests: fix xfsprogs building and promote xfsdump building as an option
To build xfsprogs on debian we need a few more packages, update them.
Also, I noticed that the original code git cloned xfsprogs as root,
so fix that by checking if the directory already exists and if so
setting the permissions appropriately.

Then, in passing I noticed we were only giving the option to build
xfsdump for debian. Promote this as a proper kconfig option and let
anyone select it. It used to be that debian stopped shipping xfsdump
but the package is back up, so we default to just install the distro
package now on debian and you have to enable the new option to build
it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>