create_partition: fix checking existing partition
If the partition already existed we fail to use the actual return code,
and can sometimes re-try to wipe the device by mistake. Fix this.

This was spotted by running 'make fstests' more than once on the
sparsefiles sparsefiles_device partition. I am not sure why this
was failing on the sparsefile partition but not on the data partition
if it already existed.. I tried several times already with the /data/
partition there and there was no issues.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>