workflows: Workflow stub for Jorge Mora's nfstest suite
The nfstest suite is a detailed test of NFS functionality for NFS
versions 3 and newer, including delegation, attribute cache
consistency, sparse files, and server-to-server copy offload.

When configured in dedicated mode, each nfstest suite is run in a
separate target node to achieve good parallelism. You can select
specific test groups to run via Kconfig options.

This commit introduces the workflow's documentation and Kconfig
apparatus to show how this workflow can be configured. Following
commits will bring in the kdevops components to get the workflow
working.

The careful reader will note that not every nfstest group is
available here. Three groups require the use of a second client,
and I haven't figured out how that needs to work. For now, those
groups are not included in this series.

Cc: Jorge Mora <Jorge.Mora@netapp.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>