common: fix setting data_user and group by default
The move to kdevops default username/group broke vagrant uses as
we were only adding the variables when the user *disabled* inference,
however there are many places which don't include the common task, and
so the default user is used in those cases. Instead of breaking them
by requiring each use to include the common role, just default the
vagrant username and group to vagrant always. If a task wants to use
the inference they can include the common role.

Reported-by: 	Chuck Lever <chuck.lever@oracle.com>
Fixes: de140b0f3eff ("common: move getting user and group inference to shared role"
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>