fstests oscheck: use a fallback expunge if one is found
Today we if we work on say Linus's v6.8-rc2 kernel and do some changes
we need to then rely on creating a v6.8-rc2+ directory symlink for
expunges if we want to rely on the same baseline. Experience shows
we just create these symlinks over and over, in fact we always just
do that. This is silly, so just use a fallback to look for the
expected directory if one is not found.

Note that this does not affect your workflow with kdevops on fstests
if you already had a symlink, that will just be used first.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>