xfs: update expunges with two new crashes one related to btrfs sparse files
One is already known, the other one seems new, and it is triggerable
by using sparse files on a btrfs partition for testing xfs on test
generic/476 (no SOAK_DURATION needed), so the bug is specific to btrfs.
The bug is triggered on the btrfs-endio-write btrfs_work_helper
workqueue and hits BUG_ON() on a dad page state on the call
btrfs_release_extent_buffer_pages() [0].

We've done extensive testing of XFS on v6.6-rc5 and recently we moved
to v6.8, and so this could be an issue introduced on v6.7 or v6.8.

[0] https://gist.github.com/mcgrof/76e026ca48b95922a365be3502fddf45

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>