terraform: fix missing nodes.tf
When we moved nodes file generation to jinja2 the commit relied on
symlinks to the main directory. This may have worked but after the
new git tree move the symlink was lost for some reason, even though
I cannot understand why. I found that bringing up guests on AWS broke
because the terraform/nodes.tf was generated but the terraform/aws/nodes.tf
was not present, so perhaps its that the .gitignore on terraform/.gitignore
prevented to carry on the old symlink. Either way, this is rather finicky
so just avoid the symlink and make the KDEVOPS_NODE dependent on the
actual cloud provider a first class declared citizen.

Fixes: f847692d7a ("terraform: provide proper generation for nodes with jinja2")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>