xfs: add xfs/242 as failing on xfs_reflink_2k
This test is rather simple, and somehow we managed to capture a
non-crash failure. The test was added to fstests via fstests commit
0c95fadc35c8e450 ("expand 252 with more corner case tests") which
essentially does this:

+       $XFS_IO_PROG $xfs_io_opt -f -c "truncate $block_size" \
+               -c "pwrite 0 $block_size" $sync_cmd \
+               -c "$zero_cmd 128 128" \
+               -c "$map_cmd -v" $testfile | $filter_cmd

The map_cmd in this case is: 'bmap -p'. So the test does:

a) truncates data to the block size
b) sync
c) zero-fills the the blocksize

The xfs_io bmap displays the block mapping for the current open file.
Since our failed delta is:

-0: [0..7]: data
+0: [0..7]: unwritten

So it would seem we somehow managed to race to write, but it never
went anywhere. I can't reproduce yet, but figured I'd put this out
there to at least acknowledge its seen at least once.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>