xfs: add generic/476 as hanging during compaction on 6.9.0-rc4
kdevops managed to trigger a new hung condition between the compaction
kthread kcompactd while migrating pages when fsstress is running on test
generic/476 on the xfs test profile xfs_nocrc_2k, ie with CRCs disabled
and with a 2k block size filesystem. The guest uses loopback block
devices for the fstests TEST_DEV, the loopback file uses sparsefiles
on a btrfs partition. The contention based on traces [0] seems to be
that we have somehow have fsstress on fallocate + compaction race on
folio_wait_bit_common().

We have going:

  a) kthread compaction --> migrate_pages_batch()
                --> folio_wait_bit_common()
  b) workqueue on btrfs writeback wb_workfn  --> extent_write_cache_pages()
                --> folio_wait_bit_common()
  c) workqueue on loopback loop_rootcg_workfn() --> filemap_fdatawrite_wbc()
                --> folio_wait_bit_common()o
  d) kthread xfsaild --> blk_mq_submit_bio() --> wbt_wait()

I don't have a failure rate yet. Given we also have korg#218227 and
a fix posted for it [2] and that issue was also induced by fsstress
and compaction running at the same time, its unclear if this is a
regression after v6.8-rc2.

[0] https://gist.github.com/mcgrof/4dfa3264f513ce6ca398414326cfab84
[1] https://bugzilla.kernel.org/show_bug.cgi?id=218227
[2] https://lore.kernel.org/all/7ee2bb8c-441a-418b-ba3a-d305f69d31c8@suse.cz/T/#u

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>