redhat: skip the "Enable the CodeReady repo" task if we're using a custom yum repofile
If we're using a custom yum repofile, there's no guarantee the CRB repo
is named the same or that there even is a CRB repo, so skip the task in
that case.  An alternative approach would be to add "ignore_errors: yes"
to that task (in which case we'd just fail later in the "Install build
dependencies" task if there's some package we can't find).

Signed-off-by: Scott Mayhew <smayhew@redhat.com>