nfsd: Provision an iSCSI initiator on the kdevops NFS server
Enable NFS exports to be placed on iSCSI LUNs. This is a common
deployment practice for pNFS block layout.

With this commit, NFSD configuration now gives the option of placing
the physical storage for exports locally (as is done currently) or
on an iSCSI LUN. Then all exported data on the NFS server will
reside either in a local volume group on the NFS server, or on an
external iSCSI LUN. Later we might want to add other storage
locations, like adding NVMe.

Note that although using a LUN can support any file system type
that resides on durable storage, only XFS supports the pNFS block

Cc: Michael Christie <michael.christie@oracle.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>