iscsi: Provision a target node to host iSCSI LUNs
I would like to test the NFS server's pNFS block layout type
implementation. That requires a set of block devices to be
accessible to both the NFS server and its clients. iSCSI is suited
to this task.

I'm setting the iSCSI target up as a separate host so that other
tests can also use iSCSI LUNs as file system storage.

Unfortunately, "iSCSI target" is the de facto term used to refer to
the system that houses physical storage that is accessible via the
iSCSI protocol. It is not related to kdevops' use of the term
"target node".

Cc: Michael Christie <michael.christie@oracle.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>