xfs: promote xfs/013 as an expunge to all test sections
xfs/013 can lead to a known crash reported through korg#218226 [0].
Dave noted this is known, and requires more work. The v6.9-rc6 release
is confirmed to still have the issue, we managed to capture its failure
on xfs_reflink_stripe_len on v6.9-rc6 as we just had not expunged it
for that test section.

Promote xfs/013 to all.txt so that we can avoid it if we want to when
wanting to try to test only tests which should not fail as its already
failing on 10 XFS test sections.

[0] https://bugzilla.kernel.org/show_bug.cgi?id=218226

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>