linux-mirror: add kdevops-results-archive
Since we split the results out to a new tree, add the results
git tree to the mirror list. I'm lazy so I'm not adding the
Kconfig options for alternative mirrors for it later. This is
all that is needed if you want to add a new mirror and don't
need to add variability.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>