linux-mirror: move kdevops mirroring to yaml
The yaml approach to mirroring is much simpler to support more
variability and growth, so use that. We can now remove the systemd
service and timer unit files as well, as they are generated for us.

This also demonstrates a way to customize the refresh rate of the
mirror. By default we use a refresh rate of 360m and a refresh on boot
of 10m. For kdevops we want this to be queried much more often for
our expunges.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>