guestfs: add support to infer host distro mirrororing optimizations
For production level uses of kdevops you want to optimize every ounce of
time, and network bandwidth. kdevops already mirrors everything we can
git clone, the next optimization is to ensure both the host and guest
use their own local mirrors for packaging. This is resolved by the
host and we don't care how, but *if* we can detect this optimization
in place, *and* if the host and guest match distros, then it makes
sense to leverage the host's sources file for the repo for the guest.

So, this optimization requires no user input. We start with debian
support, and add enough kconfig magic to let others easily add support
for other distros, all you need is the mapping file and how to detect
when you think it is safe from a host perspective for a guest to use
your sources file.

The first heuristic is a simple traceroute to check how many hops
your host uses for its sources, if its 1 chances are pretty high
the host is using a mirror for its packaging sources. This is also
distros specific as each source file differs, for debian its
/etc/apt/sources.list, and so we traceroute the first entry. Second
heuristic is determining when we can use the host sources on the
guest. For debian we enable this for now when the host is trixie and
the guest is either trixie or buster. That is, if the host is using
trixie we'll upgrade a buster guest to trixie.

We've always wanted trixie for folks working on hacking, its just we
can't use it yet...

If you only want buster, then your host probably only wants to use
buster too.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>