guestfs: add initial debian trixie support with custom URLs
debian does not yet provide an index file for virt-builder, but we
have URLS with qcow2 and raw files. Using them in guestfs is actually
not quite trivial. So we do the handy work to enable others to also
use custom URLs and build a virt-builder local source and index file
for you. All we need really, is to check the checksums.

Sadly this does not yet work, as it seems we get stuck on the grub
prompt for some reason. This needs some more investigation.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>