guestfs: set default root password
For guests we used to use vagrant and that had a default root
password. Setting up a custom one makes sense if we're doing
production stuff, but if you're just hacking on stuff on
ephemeral development guests in a private network we want to
be able to log in easily, not scrape random passwords. Use
kdevops for the root password.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>