guestfs: fix LIBVIRT_ENABLE_LARGEIO serial start count
When CONFIG_LIBVIRT_ENABLE_LARGEIO is enabled we modify the amount of
either virtio or nvme drives we use. The jinja2 template for the guest
xml file starts the serial id at kdevops1 instead of kdevops0, fix this.

This fixes using CONFIG_LIBVIRT_ENABLE_LARGEIO when enabling anything
that requires mounting one of the drives which assumes kdevops0 is
present such as 'make linux', which will mount /data/ first on the
kdevops0 drive.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>