guestfs: disable copying guest sources on upgrade
The heuristic to copy /etc/apt/sources.list in case of a hop 1 count
work well but we currently have it enabled when we have a distro match
on host and guest *and* also when we upgrade the guest up to the host's
distro. The later heuristic is enabled on GUESTFS_DEBIAN_HOST_UPDATES_GUEST
but in practice I'm already seeing some package renames like libaio1
on buster is now libaio1t64 on trixie... so we can't just do this
quite yet unless we have a specific mapping of packages for debian
testing.

Yes, we want to use debian testing, but guestfs doesn't yet use it
because we have no stupid working raw image yet for it. So let's not
break existing users on buster and guestfs. We can move away from
buster on guestfs once we either provide a solution to this package
name differences for debian or we get trixie working. Disable this
for now to not break.

What we should do is just allow to *re-use* the sources file but
modify it to rename trixie to buster for instance.

Fixes: 11e6bd0b4679d ("guestfs: add support to infer host distro mirrororing optimizations")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>