roles: add new pkg role
We strive to support rolling distros on the host but there is support
for older distros on the target nodes / guests. For now we've been using
each distro's rolling distro package names for install-deps for each
major ansible role, and each role special-cases their own renames for
older releases one way or another.

Provide a simple pkg role which all it does is set out to define a main
package name and a way for each distro sub-release to provide
alternatives.

For now this is just used by debian-testing but can easily grow so to
enable other distros to sprinkle support and simplify their own distro
package maps.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>