devconfig: enable generic target to trigger devconfig playbook
We run the devconfig playbook when we bringup nodes either on the cloud
or through virtualization. However, after bringup, if we're extending
some kdevops devconfig playboook stuff we have no way to easily re-run
it other than to try the misleading 'make bringup' again, and that is
just silly.

Provide a generic target for 'make devconfig' which will also help us
run essentially each distro's equivalent to an 'apt-get dist-upgrade'
if KDEVOPS_TRY_UPDATE_SYSTEMS is enabled.

This knocks two birds with one stone:

 1) We provide a nice easy interface to the devconfig playbook
 2) If you've enabled KDEVOPS_TRY_UPDATE_SYSTEMS you get to pick
    when to update your target nodes by just running 'make devconfig'
    willy nilly.

This will make KDEVOPS_TRY_UPDATE_SYSTEMS even more useful, which we'll
demonstrate in the next commit.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>