guestfs: fix GUESTFS_DEBIAN_COPY_HOST_SOURCES when upgrading to trixie
Whe GUESTFS_DEBIAN_COPY_HOST_SOURCES is enabled an upgrading buster
to trixie the default package was intended to be the one on the rolling
distro. Fix this. The pkg role renames this to the buster respective
package already for you, so we just need the default set to the trixie
version.

Fixes: b804bd105051f ("guestfs: re-enable debian buster to upgrade to trixie")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>