guestfs: fix debian trixie support
First order of business to fix was the template image,
Scott Mayhew noted we needed to enable UEFI boot, and we
can just use a simple automatic UEFI firmware and secure boot
environment.

Next issue was networking, the dhcp package and scripts to
set up the network interface are needed. Finally, the last
step was to ensure we don't get prompted when we reconfigure
the openssh-server.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>