guestfs: fix forcing uefi
Daniel reported an issue with bringup on buster, I root caused
it to the change recently made to force uefi on all guests. Revert
only that part of the commit so to ensure buster keeps working.

We'll later add support for a NEEDS_UEFI kconfig option and make
trixie select it.

Fixes: 9a207c8e0d ("guestfs: fix debian trixie support")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>