guestfs: fix uefi support
Fix trixie support. Make the UEFI requirement optional and fix ssh hangs
on buster. The issue with the hang on ssh on buster is the forced dpkg
reconfigure of ssh is it can take a while and ssh trying to restart in
the background will eventually give up and fail. Instead just stop ssh
first before running the dpkg-reconfigure. Then start ssh.

Fixes: 9a207c8e0d ("guestfs: fix debian trixie support")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>