linux: fix ref name generated in default mode
When using the automatic default generation path, the reference name
parsed from releases.json does not include the 'v' prefix (e.g., 6.10
instead of v6.10). This causes a git-clone failure when the bootlinux
playbook is run because an invalid reference is generated. Therefore,
ensure that the 'v' prefix is included when parsing the version from
releases.json when necessary.

Fixes: 47f2275 ("linux: generate refs automatically")

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>