devconfig: add chicken-and egg debian mirror workaround
When as host is detected to use a close nearby mirror (we count hops)
we enable infer that if you use virtualization you ideally want to use
the same mirror if the distributions align both on the host and guest.

A snag we sometimes run into is a chicken and egg problem where the
guest comes up, and we apt-get update but this fails as the guest
is not yet configured with NTP so their idea of time is completely
off.

To fix this, since we are working on kernel hacking and not the new york
stock exchange, fix it by adding a work around for the first part of the
devconfig playbook which updates the mirror used for apt on the guest
with:

   apt-get -o Acquire::Check-Valid-Until=false -o Acquire::Check-Date=false

This fixes the errors sometimes encountered when using mirroring on the
guests such as:

STDOUT:

Get:1 http://private-mirror.local:8081/artifactory/yay testing InRelease [169 kB]
Get:2 http://private-mirror.local:8081/artifactory/yay testing-updates InRelease [49.6 kB]
Reading package lists...

STDERR:

E: Release file for <...>dists/testing/InRelease is not valid yet (invalid for another 4h 9min 22s). Updates for this repository will not be applied.
E: Release file for <...>dists/testing-updates/InRelease is not valid yet (invalid for another 4h 9min 22s). Updates for this repository will not be applied.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>