scripts/honey-badger.py: add new stable kernel scraper like we don't care
Ever need to just tons of stable kernels fast and you just
want to build them? The Ubuntu stable kernels on their PPA list
have existed for decades. Let's just use that even if you don't
have debian based distros we can just extract the debian files using
ar for you. Just honey badge them.

This was inspired by my old use of the Ubuntu PPA kernels when I helped
with the Linux kernel backports project, which required testing on tons
of stable kernels, in it was a script backports-update-manager [0]. That
code just has tons of crap we don't need for regular kernel development,
so just remove it, and clean it up, and enhance it so we can just get
the functionality we need to install on the system.

This was also a Proof Of Concept work to see how easily and quickly
something like ChatGPT could help simplify an existing Python script
and adapt it fast for our needs on kdevops.

Will we use this? Eventually, we may, but more immediate there was a
need to get tons of stable kernels and this was just a pain in the rear
to install. This should help future testing when that need arises.

Notes:

- just update-grub or your equivalent after using it. Patches welcomed
  to do that for folks.

- adding --skip-kernels-above should be easy as a future enhancement
  which ChatGPT can kick out in no time. That would allow you use
  --skip-kernels-above v6.9 and so that will not install kernels later
  than v6.9.

[0] https://git.kernel.org/pub/scm/linux/kernel/git/backports/backports.git/tree/devel/backports-update-manager

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>