linux-mirror: complete ansible conversion
The linux-mirror effort started as a set of bash scripts [0] intended
to be used elsewhere, however, we've learned over time its best to
just define your requirements on a yaml file and process things with
ansible task. This started out a while ago with mirrors.yaml template
file we generate, and then use to generate the respective systemd
service / timer and git service and socket files. Now that we've
done the full conversion over to the yaml template the last thing
that was missing was the git service and socket file. Move that to
be installed in the ansible playbook and also provide support for
the check status via ansible as well:

  make mirror-check

The git subtree is now irrelevant, so just remove its references from
the Makefile.subtrees. Yes, perhaps we can make this easily accessible
outside of kdevops, but that's a bit outside of the scope right now,
given it should be *super easy* to just use kdevops to start mirroring.
Using kdevops to start mirroring is pretty light-weight already, so
we could just have a defconfig for that, and in fact that's how most of
us setup new kdevops systems with new mirror setups: we just configure
to enable mirroring and just run:

  make defconfig-mirror
  make
  make mirror

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>