linux-mirror: use 3 passes on start-mirroring.py
start-mirroring.py is the main entry to start processing
our mirrors.yaml when we want to start mirroring, if mirroring
is enabled. Right now we we simply iterate over all mirrors and
start mirroring on each one. This script already has reference
support but it then means we put the onus on ordering of
mirrors.yaml, which also means then that alphabetical ordering
of the file is not possible.

Instead of putting the onus of ordering by reference dependencies
on mirrors.yaml, just have the processing of mirrors.yaml be split
up in 3 passes:

 1) verification that all entries are correct
 2) mirror entries which do not have reference
 3) mirror entries which require a reference

We don't have trees that have double reference, so a--> b--> c,
we just have a--> b, in this model the above strategy works fine
and suffices. This let's us keep alphabetical ordering on
mirrors.yaml.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>