linux-mirror: move linux-stable.git mirror to yaml
This moves the linux-stable mirror to the yaml entry, therefore
simplifying it with a simple yaml description.

The playbooks/roles/linux-mirror/linux-mirror-systemd/Makefile style
of declaring mirrors is being replaced with the more simple yaml
description. This let's us autogerate the systemd unit and mirror
files from a simple yaml description file.

The diff stat says it all:

9 files changed, 24 insertions(+), 36 deletions(-)

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>