linux-mirror: enable support to mirror forked trees using an org as a directory
We mirror the latest xfstests-dev.git tree but we also have at times
a delta we want to carry of that same tree under kdevops/fstests.git,
in such case /mirror/xfstests-dev.git may exist and we want for
convenience to also have /mirror/kdevops/fstests.git. To support this
we allow the yaml target to use this convention "kdevops/fstests.git"
but we just need to replace "/" with "-" when creating the respective
systemd unit service and timer files.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>