gitr: Make gitr work on RHEL/CentOS
RHEL and CentOS do not ship cvsps or perl-TAP-Harness-Archive.  For the
former, we'll have to build it from git.  For the latter, we can install
it from CPAN, but we'll also have to install the perl-App-cpanminus
package to do so.

Also add 'make' and 'tar' to the dependencies list.  These were
previously getting installed inadvertently by the "Install btrfs-progs"
task in playbooks/roles/devconfig/tasks/install-deps/redhat/main.yml,
which was installing more than just btrfs-progs.

Reviewed-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Scott Mayhew <smayhew@redhat.com>