scripts/kconfig/kconfig.Makefile: move _refs-default
We should avoid editing scripts/kconfig as much as possible unless
you are aware of the git subtree management for kconfig. So move
the old change to depend on _refs-default over to the Kconfig
dependency we had added to allow dynamic Kconfig. In the old
case we only enabled dynamic kconfigs when someone ran 'make dynconfig'
but Dan's patches which added ref tag support lets us now use dynamic
kconfig always but with minimal things: kernel branches for static
references defined in-tree on yaml files. That is we don't talk online
to fetch information.

Move the _refs-default dependency to the Kconfig dependency to make
this more explicit and to keep scripts/kconfig clean so to allow
easier updates from git subtrees.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>