kdevops: remove duplicate symbols and use scripts/append-makefile-vars.sh
A few kdevops kconfig duplicate symbols exist, remove them so to ensure
we can work with the latest linux-next version of kconfig. Additionally
we need to ensure to use scripts/append-makefile-vars.sh for when we
want to work with Makefile inherited variables. The KDEVOPS_PLAYBOOK_DIR
can also just be simplified, while we do provide for variability in
practice we never use this so just this a silent default for now.

Without this we end up in a hung state if we run 'make menuconfig'
with the latest kconfig. Fix this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>