scripts/ssh.Makefile: remove Makefile appending of KDEVOPS_HOSTS_PREFIX
We have CONFIG_KDEVOPS_HOSTS_PREFIX but have two respective ansible
variable:

  * kdevops_host_prefix
  * kdevops_hosts_prefix

We need to keep these for now to ensure old configs work and so existing
extra_vars.yaml continue to work. We at least annotate these for future
unification into a release where folks are not going to use the old
symbols on existing running environments.

We can later look for these with:

  git grep ^"# old-yamlconfig-compat"

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>