fstests/nfs: add ability to test with NFS loopback mounts
Add the ability to test with NFS loopback mounts (meaning NFS mounts
where the NFS server is also a client - not necessarily using the
loopback interface).  The primary motivation is to facilitate testing of
NFS with LOCALIO support:

https://lore.kernel.org/linux-nfs/0bd8f46b4b4c1e0bc9008014e8442fa3a7f3afe5.camel@kernel.org/T/#t

The changes should be pretty straightforward for the most part, the one
exception being the change to the "Check which fstests test types are
enabled" task in both the gen_hosts and gen_nodes playbooks.  Here we
add some logic to cause the "[nfsd]" section of nfs.config to be
ignored.  The reason for that is so that we don't wind up with a value
of "nfs-nfsd" to be added fstests_enabled_test_types list, which would
then cause a bogus "{{ kdevops-hosts-prefix }}-nfs-nfsd guestfs node
configuration to be created as well as a bogus entry to be added to the
ansible inventory file.

Reviewed-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Scott Mayhew <smayhew@redhat.com>