ci: add initial github / gitlab workflow tests scripts
These are extremely bare bones but with good reason on my analysis
of github actions and gitlab pipelines:

 - github has decided they think everyone on the planet is OK with
   an ubuntu based container to run the world
 - clearly the above assumption makes github completely useless
 - the gitlab runners which can run on vms need much more special
   effort
 - both rely on containers anyway
 - containers assume no systemd
 - kdevops heavily relies on systemd
 - we need selinux / apparmor disabled

And so as-is both github and gitlab are just useless. So we can use
them for the basic stuff. And even then just make mrproper works.

To do some real testing we'll need dedicated runners per distribution.
But with this we can start extending basic sanity tests.

Both of these CI tests are successful. Please don't break them, even
though they are stupid simple. We should strive to at least get
make menuconfig / nconfig / and maybe a defconfig and make going somehow
but that may take some time given the heuristics we have for Kconfig and
you running a real Linux distribution, not some child Fisher Price Github
concept of what Linux is.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>