CI: add initial self-hosted runner: fstests
This adds our very first real VM test which leverages the self
hosted github runners we assign to kdevops. We start off with a
simple workflow to ensure we can do:

1) make
2) make bringup
3) make linux
4) make fstests

We can grow these tests but if you do, consider adding at least
a self-hosted runner. We can try to scale this later with cloud
providers as well.

We should at least now never regress the fstests workflow.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>