sysbench: add TPS variability plots
Allow us to plot sysbench TPS variability so we can visualize results.
This plots TPS for each host and then if we are doing A Vs B testing we
plot TPS of A vs B. The hard part was the TPS variability stuff which
has its own script and tons of different ways to visualize this. We
define the TPS variability as the square of the standard deviation.

The TPS outliers are also plotted and they are just cattered randomly
outside 1.5 times the IQR range [0].

[0] https://en.wikipedia.org/wiki/Interquartile_range

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>