sefltests: add support for optional timeout and process TAP output
Kernel selftests are inteneded to be rather "fast" [0], and so the
default timeout is 45 seconds. And although some test can override
the timeout, we don't want this to grow out of hand. However even if
tests are "rather fast" test runners may need more time, and this will
depend on the configuration of the test runner.

kdevops has support for *how* we configure a test runner, and so we
know what we are using. For the default virtualization setup we use
and for defualt cloud setups we have, 165 seconds should suffice for now
for all kmod tests, so use that.

Then the last missing piece was a way to process TAP output and fail
in case of an actual test failure. To do that we just look for "not ok"
on the TAP output for each test. This should allow us to start using git
runners for selftests for kdevops now!

[0] https://www.kernel.org/doc/Documentation/dev-tools/kselftest.rst

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>