linux-mirror: simplify defconfig and add NFS
It actually is a much better idea to setup the mirror as root,
and so while at it augment the defconfig with NFS support as
desirable. We have a check in place early on the linux-mirror
playbook which will check for the root user early and bail if
not.

We can also simplify the defconfig by making it clear we are going
to skip bringup and we not enabling any workflow at all.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>