linux-mirror: install git-daemon when using NFS for /mirror
When you use an NFS server for /mirror/ we still need to enable
our guests the ability to connect to git. To do this we use the
git-daemon. The only issue I've found is that DynamicUser=yes
needs to be disabled on *some* systems. In this case I've determined
this is an issue with the latest and greatest debian testing but
a few months ago this was not an issue. So if you are using and
old run you won't have issues. Once you upgrade though you'll run
into issues so best to just disable it.

This adds the necessary checks to infer if you have an NFS server
set up for /mirror/ and disables the options to install a git setup
on /mirror/, but we will allow you to setup a git-daemon for you.

This is designed to also work well with:

make defconfig-mirror

So if you want to set up a kdevops mirror you first should run it
as root and use:

make defconfig-mirror
make -j80

Once you have the mount setup on hosts in the network, kdevops will
detect this in the new kconfig variable LINUX_MIRROR_MOUNTED_ON_NFS
and if this is enabled we default to only allowing to instal the
git daemon setup for you. So you can rely on the same exact defconfig
on hosts where you just want to setup the git daemon then too:

make defconfig-mirror
make -j80

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>