selftests: fix xarray and maple tree testing
The initial commit to add xarray and maple tree was pretty sloppy as
it assumed that they were selftests. They are not. And so we are going
to leverage the "selftests" idea of parallelizing special tests like
for xarray and maple and allow them to also be run in another guest but
they are just not proper kernel sefltests. They *could* be, but that
is a seprate topic and future kernel commit.

So for now we just leverage the selftests kdevops infrastructure for
spawning two guests for each tests. Both xarray and the maple tree are
special in that they both can be tested with just plain userspace and
also with a kernel module. We add support for both.

We collect the output for both as well.

To support all this I had to add heuristics to determine if a test is
a proper kernel selftest or not. This should be good to go for our
kernel-ci!

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>